Comp methods that run when the parent pawn crafts something.
Added a patch to GenRecipe.MakeRecipeProducts that runs whenever something is crafted. Added a Notify_ThingCrafted extension method to HediffWithComps and ThingWithComps that is run by the aforementioned patch. Added a ThingComp and a HediffComp with a Notify_ThingCrafted method that is run by the Notify_ThingCrafted extensions. Currently untested and undocumented.